style: add skeleton laoder for subnet selection
